Adding test for DeletionPolicy flag in transform.config.json. Renaming all projects to lowercase without dashes to make CLI happy.

+++ packages/amplify-e2e-tests/__tests__/api.test.ts
@@ -5,8 +5,10 @@ import {
   amplifyPush,
   amplifyPushUpdate
 } from '../src/init';
+import * as path from 'path';
+import { existsSync } from 'fs';
 import { addApiWithSchema, updateApiSchema } from '../src/categories/api';
-import { createNewProjectDir, deleteProjectDir, getProjectMeta } from '../src/utils';
+import { createNewProjectDir, deleteProjectDir, getProjectMeta, updateConfig, getTable, deleteTable } from '../src/utils';
 
 describe('amplify add api', () => {
   let projRoot: string;
@@ -16,7 +18,10 @@ describe('amplify add api', () => {
   });
 
   afterEach(async () => {
-    await deleteProject(projRoot);
+    const metaFilePath = path.join(projRoot, 'amplify', '#current-cloud-backend', 'amplify-meta.json');
+    if (existsSync(metaFilePath)) {
+      await deleteProject(projRoot);
+    }
     deleteProjectDir(projRoot);
   });
 
@@ -49,5 +54,34 @@ describe('amplify add api', () => {
     await expect(GraphQLAPIIdOutput).toBeDefined()
     await expect(GraphQLAPIEndpointOutput).toBeDefined()
     await expect(GraphQLAPIKeyOutput).toBeDefined()
-  })
+  });
+
+  it('inits a project with a simple model and then migrates the api', async () => {
+    const projectName = 'retaintables';
+    const initialSchema = 'simple_model.graphql';
+    console.log(projRoot);
+    await initProjectWithProfile(projRoot, { name: projectName });
+    await addApiWithSchema(projRoot, initialSchema);
+    updateConfig(projRoot, projectName, {
+      TransformerOptions: {
+        '@model': { EnableDeletionProtection: true }
+      }
+    });
+    await amplifyPush(projRoot);
+    const projectMeta = getProjectMeta(projRoot);
+    const region = projectMeta.providers.awscloudformation.Region;
+    const { output } = projectMeta.api[projectName];
+    const { GraphQLAPIIdOutput, GraphQLAPIEndpointOutput, GraphQLAPIKeyOutput } = output;
+    await expect(GraphQLAPIIdOutput).toBeDefined()
+    await expect(GraphQLAPIEndpointOutput).toBeDefined()
+    await expect(GraphQLAPIKeyOutput).toBeDefined()
+    await deleteProject(projRoot);
+    const tableName = `Todo-${GraphQLAPIIdOutput}-integtest`;
+    const table = await getTable(tableName, region);
+    expect(table.Table).toBeDefined()
+    if (table.Table) {
+      const del = await deleteTable(tableName, region);
+      expect(del.TableDescription).toBeDefined()
+    }
+  });
 });
